+++ src/Infrastructure/Filesystem/Base64Encoder.php
@@ -9,12 +9,43 @@ use Symfony\Component\Filesystem\Filesystem;
 
 readonly class Base64Encoder
 {
+    private array $allowedRecordFileExtensions;
+
     public function __construct(
         private Filesystem                                    $filesystem,
         private \Symfony\Component\Mime\Encoder\Base64Encoder $base64Encoder,
         private LoggerInterface                               $log
     )
     {
+        $this->allowedRecordFileExtensions = ['wav', 'mp3'];
+    }
+
+    /**
+     * @param non-empty-string $filename
+     * @throws InvalidArgumentException
+     * @throws FileNotFoundException
+     */
+    public function encodeCallRecord(string $filename): string
+    {
+        if (!$this->filesystem->exists($filename)) {
+            throw new FileNotFoundException(sprintf('file %s not found', $filename));
+        }
+
+        $fileExt = pathinfo($filename, PATHINFO_EXTENSION);
+        if (!in_array($fileExt, $this->allowedRecordFileExtensions, true)) {
+            throw new InvalidArgumentException(sprintf('wrong record file extension %s, allowed types %s',
+                $fileExt, implode(',', $this->allowedRecordFileExtensions)));
+        }
+
+        $fileBody = file_get_contents($filename);
+        if (false === $fileBody) {
+            throw new InvalidArgumentException(sprintf('cannot read file %s', $filename));
+        }
+
+        $fileBody = $this->base64Encoder->encodeString($fileBody);
+
+        $this->log->debug('encodeFile.finish');
+        return $fileBody;
     }
 
     /**
@@ -35,7 +66,7 @@ readonly class Base64Encoder
 
         $fileBody = $this->base64Encoder->encodeString($fileBody);
 
-        $this->log->debug('encodeFile.finish¨');
+        $this->log->debug('encodeFile.finish');
         return $fileBody;
     }
 }